Added authorization for loans, payments and users

+++ src/lib/controller/LoanController.php
@@ -35,9 +35,19 @@ class LoanController {
 	}
 
 	public function getLoanHandler(Request $req, Response $res): Response {
-
+		
 		try {
 			$loan = \Uomi\Model\Loan::findOrFail( $req->getAttribute('loan_id') );
+
+			$auth = new Authentication();
+			$isTo = $auth->isRequestAuthorized($req, $loan->to_user);
+			$auth = new Authentication();
+			$isFrom = $auth->isRequestAuthorized($req, $loan->from_user);
+
+			if(!($isTo || $isFrom)) {
+				return $auth->unathroizedResponse($res, $auth->getErrors());
+			}
+
 			$stat = new Status($loan);
 			$stat = $stat->message("Loan found");
 			return $res->withStatus(200)->withJson($stat);
@@ -80,6 +90,14 @@ class LoanController {
 
 		try {
 			$loan = \Uomi\Model\Loan::findOrFail( $req->getAttribute('loan_id') );
+
+			$auth = new Authentication();
+			$isFrom = $auth->isRequestAuthorized($req, $loan->from_user);
+
+			if(!($isFrom)) {
+				return $auth->unathroizedResponse($res, $auth->getErrors());
+			}
+
 			$loan->details = $details;
 			$loan->category_id = $catModel->id;
 			$loan->save();
@@ -102,6 +120,16 @@ class LoanController {
 		$from_user = $form['from_user'] ?? null;
 		$amount_cents = $form['amount_cents'] ?? null;
 		$category_id = $form['category_id'] ?? null;
+
+
+		$auth = new Authentication();
+		$isTo = $auth->isRequestAuthorized($req, $to_user);
+		$auth = new Authentication();
+		$isFrom = $auth->isRequestAuthorized($req, $from_user);
+
+		if(!($isTo || $isFrom)) {
+			return $auth->unathroizedResponse($res, $auth->getErrors());
+		}
 
 
 		if($to_user === null) {
@@ -150,6 +178,16 @@ class LoanController {
 
 		try {
 			$loan = \Uomi\Model\Loan::findOrFail( $req->getAttribute('loan_id') );
+
+
+			$auth = new Authentication();
+			$isFrom = $auth->isRequestAuthorized($req, $loan->from_user);
+
+			if(!($isFrom)) {
+				return $auth->unathroizedResponse($res, $auth->getErrors());
+			}
+
+
 			$loan->delete();
 			$stat = new Status();
 			$stat = $stat->message("Loan deleted");
